feat(category): 实现分类删除功能并优化唯一性校验

+++ only-danmuku-adapter/src/main/kotlin/edu/only4/danmuku/adapter/portal/api/AdminCategoryController.kt
@@ -95,7 +95,7 @@ class AdminCategoryController {
     fun adminCategoryDel(@RequestBody @Validated request: AdminCategoryDel.Request): AdminCategoryDel.Response {
         Mediator.commands.send(
             DeleteCategoryCmd.Request(
-                categoryId = request.categoryId.toLong()
+                categoryId = request.categoryId
             )
         )
         return AdminCategoryDel.Response()
